Ajout de test pour tournoi

+++ Client2Mongo.py
@@ -31,23 +31,39 @@ class Client2Mongo:
         niveau = tournoi.categorie[1]
 
         doc = {"nom": nom, "date": {"debut": date, "fin": date}, "format": format,
-               "categorie": {"age": age, "niveau": niveau}, "status": "Prevu"}
+               "categorie": {"age": str(age[0])+"-"+str(age[1]), "niveau": niveau}, "status": "Prevu"}
         tournois.insert_one(doc)
 
     def modif_tournoi(self, old_nom_tourn, new_nom_tourn):
         tournois = self.get_collection("tournois")
-        filtre = {"nom": old_nom_tourn}
-        nouvelles_valeurs = {"$set": {"nom": new_nom_tourn}}
-        tournois.update_one(filtre, nouvelles_valeurs)
+        tournoi = tournois.find_one({"nom": old_nom_tourn})
+
+        if tournoi is None:
+            print("Aucun tournoi n'a été trouvé avec ce nom : ", old_nom_tourn)
+        else:
+            filtre = {"nom": old_nom_tourn}
+            nouvelles_valeurs = {"$set": {"nom": new_nom_tourn}}
+            tournois.update_one(filtre, nouvelles_valeurs)
 
     def affiche_tournois(self):
         tournois = self.get_collection("tournois")
         for tournoi in tournois.find():
             print(tournoi)
 
-    def affiche_tournoi(self):
-        pass
+    def affiche_tournoi(self, nom_tournoi):
+        tournois = self.get_collection("tournois")
+        tournoi = tournois.find_one({"nom": nom_tournoi})
+
+        if tournoi is None:
+            print("Aucun tournoi n'a été trouvé avec ce nom : ", nom_tournoi)
+        else:
+            print(tournoi)
 
     def suppresion_tournoi(self, nom_tournoi):
         tournois = self.get_collection("tournois")
-        tournois.delete_one({"nom": nom_tournoi})
+        tournoi = tournois.find_one({"nom": nom_tournoi})
+        if tournoi is None:
+            print("Aucun tournoi n'a été trouvé avec ce nom : ", nom_tournoi)
+        else:
+            print("Supression du tournoi : ", nom_tournoi)
+            tournois.delete_one({"nom": nom_tournoi})
